Upgrade scenario testing changes
This adds series and openstack upgrade scenario tests (the equivalent of
the existing mojo tests) for focal, plus the start of the series upgrade
replacement tests for trusty->xenial and xenial->bionic.

tox.ini is upgraded to pin to < 20.3 to avoid the new pip resolver until
that piece of tech debt is resolved.

chardet is pinned to 3.0.4 due to some installation issues on xenial.
This is again tech debt that will be resolved with the pip resolver
work.