Stop pinning the etcd charm
We were pinning an old version (474 where the latest
is 607). The git history shows no particular reason
for this (it is there from day one).

I'm seeing instabilities in this charm's layer-snap
where `snap refresh --list` would lead to a hook
error when getting an HTTP/503 from the store.
Trying out the latest version to see if this helps.