Distro test enhancements
There are several enhancements made here for distro testing.

All of the existing charms and features that were tested prior to
this change continue to be tested. Bundles have been expanded to
include testing of more networking features and to test on MAAS.
The bundles are broken down as follows to test various features:
  - swift (minimal with keystone since other bundles use ceph)
  - ovs (no-l3ha+no-dvr)
  - ovs (l3ha+dvr)
  - ovs (l3ha+dvr-snat)
  - ovs (sriov)
  - ovs (dpdk) (not yet implemented)
  - ovn (no-sriov)
  - ovn (sriov)
  - ovn (dpdk) (not yet implemented)

New zaza-openstack-tests support is used in order to:
  - reboot SR-IOV machines (required to finalize installation)
  - setup VLAN provider overcloud network for SR-IOV
  - run tempest before/after upgrading to proposed pocket or ppa