Report proposed packages per unit
For SRUs to Ubuntu and the Cloud Archive, bugs need to be updated
with testing results and package versions tested. This will enableo
getting the package versions from charm units in an automated way.

This is done prior to running tempest tests to ensure packages
are reported regardless of success/failure of tempest tests.